+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.14.0] - 2026-02-22
 ### Added
  - Use `$STEAM_COMPAT_DATA_PATH/pfx` as the game's prefix path if `STEAM_COMPAT_DATA_PATH` environment variable is set
  - Add `--list-all/-L` command to list all games for all found Steam installations
@@ -337,7 +337,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of Protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.13.1...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.14.0...HEAD
+[1.14.0]: https://github.com/Matoking/protontricks/compare/1.13.1...1.14.0
 [1.13.1]: https://github.com/Matoking/protontricks/compare/1.13.0...1.13.1
 [1.13.0]: https://github.com/Matoking/protontricks/compare/1.12.1...1.13.0
 [1.12.1]: https://github.com/Matoking/protontricks/compare/1.12.0...1.12.1